Removing duplicate code from ExecutionController and ExecutorManager. Adding it in AbstractExecutorManagerAdaptor.
